Conflicts:
	source/AccountExecutiveApp/AccountExecutiveApp.iOS/AccountExecutiveApp.iOS.csproj
	source/Database.MatchGuide/Scripts/011 - Add Users.sql
	source/Shared.Core/IMatchGuideApi.cs
	source/Shared.Core/MatchGuideApi.cs

+++ source/ConsultantApp/ConsultantApp.iOS/Remittances/RemittanceCell.cs
@@ -16,7 +16,7 @@ namespace ConsultantApp.iOS
 
 		}
 
-		public void UpdateCell(string depositDate, string documentNumber, float amount, string period)
+		public void UpdateCell(string depositDate, string documentNumber, decimal amount, string period)
 		{
 			depositDateLabel.Text = depositDate;
 			documentNumberLabel.Text = documentNumber;
